+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0]
+
+### Changed
+
+    - Code refactoring.
+
+### Added
+
+    - New tasks : 
+      - Run app in Speculos,
+      - Kill Speculos,
+      - On device functional tests,
+      - Device quick onboarding.
+
 ## [0.0.6]
 
 ### Fixed